TC: Upload de PDF para Slide (substitui o postman)

+++ Server/server.js
@@ -1,15 +1,12 @@
-// Using Express to create a server that handles file uploads and PDF conversion
-const express = require('express');
-const app = express();
 
-
-// Handling CORS issues
-const cors = require('cors');
-app.use(cors());
-app.use(express.static('public')); // Serve static files from the 'public' directory
-const PORT = 3000;                  // Port number for the server
-const fs = require('fs');           // Using File System (fs) to handle file operations
-const path = require('path');       // Get the current working directory
+const express = require('express');     // Using Express to create a server that handles file uploads and PDF conversion
+const cors = require('cors');           // Handling CORS issues
+const app = express();                  // Create an instance of Express
+app.use(cors());                        // Enable CORS for all routes
+app.use(express.static('public'));      // Serve static files from the 'public' directory
+const PORT = 3000;                      // Port number for the server
+const fs = require('fs');               // Using File System (fs) to handle file operations
+const path = require('path');           // Get the current working directory
 
 // Using Multer for handling multipart/form-data, which is used for uploading files
 // Multer is a middleware for handling multipart/form-data, which is primarily used for uploading files
@@ -26,12 +23,39 @@ app.listen(PORT, () => console.log(`Server listening on port: ${PORT}`));
 // Endpoint to handle file uploads for target files
 // The target files are expected to be in .mind format
 app.post('/upload/target', upload.single('targets'), (req, res) => {
+
+    // Log the request body and file information for debugging
+    console.log('Request body on /upload/target:', req.body);
+    console.log('File information:', req.file);
+
+    // Check if a file was uploaded
     if (!req.file) return res.status(400).send('No file uploaded.');
-  
-    const randomId = Math.floor(Math.random() * 1000000) + 1;
-    const fileName = `${randomId}.mind`; // You can use .txt, .bin, or .mind
+
+    // const randomId = Math.floor(Math.random() * 1000000) + 1;
+    // const fileName = `${randomId}.mind`; // You can use .txt, .bin, or .mind
+    // const filePath = path.join(__dirname, 'public', fileName);
+
+    // Verify if codetgt is available, if not, generate a random ID
+    let codetgt = req.body.codetgt;
+    if (!codetgt) {
+      // Generate a random ID for the target file name
+      codetgt = Math.floor(Math.random() * 1000000) + 1;
+    }
+
+    // Get the file name and path for the uploaded target file
+    const fileName = `${codetgt}.mind`; // Use the codetgt parameter from the request body
     const filePath = path.join(__dirname, 'public', fileName);
-  
+
+    // Verify if the file name is valid (not empty)
+    if (!fileName) {
+      return res.status(400).send('Invalid file name.');
+    }
+
+    // Check if the file name already exists
+    if (fs.existsSync(filePath)) {
+      return res.status(400).send('File already exists. Please choose a different name.');
+    }
+
     // Save the uploaded file to the server
     fs.writeFile(filePath, req.file.buffer, err => {
       if (err) {
@@ -39,156 +63,162 @@ app.post('/upload/target', upload.single('targets'), (req, res) => {
         return res.status(500).send('Failed to save file.');
       }
       console.log(`File saved as ${fileName}`);
-      res.status(200).send({ filename: fileName });
+      res.status(200).send({ fileName: fileName });
     });
   });
 
 // Endpoint to handle file uploads for slides (PDF files)
 // The slides are expected to be in PDF format
 // The PDF files are converted to images (PNG) and an HTML file is generated for A-Frame
-app.post('/upload/slides', upload.single('slides'), async (req, res) => {  
-    if (!req.file) return res.status(400).send('No file uploaded.');
-    // Ensure codetgt is available (modify according to your data source)
-    if (!req.body.codetgt) {
-      return res.status(400).send('Missing codetgt parameter');
+app.post('/upload/slides', upload.single('slides'), async (req, res) => 
+{  
+  // Log the request body and file information for debugging
+  console.log('Request body on /upload/slides:', req.body);
+  console.log('File information:', req.file);
+
+  if (!req.file) return res.status(400).send('No file uploaded.');
+  // Ensure codetgt is available (modify according to your data source)
+  if (!req.body.codetgt) {
+    return res.status(400).send('Missing codetgt parameter');
+  }
+
+  // Get the file name and path for the uploaded PDF file
+  // The file name is generated based on the codetgt parameter (codetgt) provided by the user
+  const fileName =  `${req.body.codetgt}` + `.pdf`;
+
+  // Join the current directory with the public directory and the file name
+  const filePath = path.join(__dirname, 'public', fileName);
+
+  // Save the uploaded file to the server
+  // The file is saved in the public directory with the name given by the user
+  fs.writeFile(filePath, req.file.buffer, err => {
+    if (err) {
+      console.error('Error saving file:', err);
+      return res.status(500).send('Failed to save file.');
     }
-    
-    // Get the file name and path for the uploaded PDF file
-    // The file name is generated based on the codetgt parameter (codetgt) provided by the user
-    const fileName =  `${req.body.codetgt}` + `.pdf`;
+  });
 
-    // Join the current directory with the public directory and the file name
-    const filePath = path.join(__dirname, 'public', fileName);
+  //const pdfPath = req.file.path;
+  const outputDir = path.join(__dirname, 'public');
+  const baseFileName = req.body.codetgt;
 
-    // Save the uploaded file to the server
-    // The file is saved in the public directory with the name given by the user
-    fs.writeFile(filePath, req.file.buffer, err => {
-      if (err) {
-        console.error('Error saving file:', err);
-        return res.status(500).send('Failed to save file.');
-      }
-    });
+  // PDF conversion options
+  const convertOptions = {
+    density: 100,
+    saveFilename: baseFileName, // Base name for output files
+    savePath: outputDir,        // Output directory
+    format: 'png',
+    width: 800,
+    height: 600,
+    compression: 'jpeg',       // Optional compression
+    quality: 90                // Image quality (1-100)
+  };
 
-    //const pdfPath = req.file.path;
-    const outputDir = path.join(__dirname, 'public');
-    const baseFileName = req.body.codetgt;
+  // Convert PDF to images
+  const convert = fromPath(filePath, convertOptions);
+  //Legacy code to convert all pages recieveing the number of pages from the client
+  // for(let i = 1; i<=req.body.numpags; i++){
+  //   const images = await convert(i);
+  // }
 
-    // PDF conversion options
-    const convertOptions = {
-      density: 100,
-      saveFilename: baseFileName, // Base name for output files
-      savePath: outputDir,        // Output directory
-      format: 'png',
-      width: 800,
-      height: 600,
-      compression: 'jpeg',       // Optional compression
-      quality: 90                // Image quality (1-100)
-    };
-
-    // Convert PDF to images
-    const convert = fromPath(filePath, convertOptions);
-    //Legacy code to convert all pages recieveing the number of pages from the client
-    // for(let i = 1; i<=req.body.numpags; i++){
-    //   const images = await convert(i);
-    // }
-    
-    // Start converting pages one by one
-    // and stop when an error occurs (e.g., no more pages)
-    let page = 1;         // Start from the first page
-    const maxPages = 200; // Set a maximum number of pages to avoid infinite loops
-    while(page <= maxPages){
-      // Convert the current page
-      try
-      {
-        // Convert the current page and save it as a PNG file
-        await convert(page);
-        page++;
-      }
-      // If an error occurs, it usually means there are no more pages to convert or the page number is invalid
-      // In this case, we can break the loop
-      catch(err)
-      {
-        console.log(`🛑 Stopped at page ${page - 1}:`, err.message || err);
-        break;
-      }
+  // Start converting pages one by one
+  // and stop when an error occurs (e.g., no more pages)
+  let page = 1;         // Start from the first page
+  const maxPages = 200; // Set a maximum number of pages to avoid infinite loops
+  while(page <= maxPages){
+    // Convert the current page
+    try
+    {
+      // Convert the current page and save it as a PNG file
+      await convert(page);
+      page++;
     }
+    // If an error occurs, it usually means there are no more pages to convert or the page number is invalid
+    // In this case, we can break the loop
+    catch(err)
+    {
+      console.log(`🛑 Stopped at page ${page - 1}:`, err.message || err);
+      break;
+    }
+  }
 
-    // Delete the original PDF file after conversion
-    fs.rm(filePath, { recursive: true, force: true }, (err) => {
-      if (err) console.error('Error deleting file:', err);
-    } );
-    
+  console.log(`PDF converted to images: ${page - 1} pages processed.`);
 
-    // Generate HTML code for the A-Frame scene
-    // This code creates a scene with a camera and an image target
-    let gencode = `
-    <a-scene id="example-target" mindar-image="imageTargetSrc: http://localhost:3000/${req.body.codetgt}.mind; uiScanning:yes; autoStart: false;" color-space="sRGB" renderer="colorManagement: true" vr-mode-ui="enabled: false" device-orientation-permission-ui="enabled: false">
-      <a-camera position="0 0 0" look-controls="enabled: false" cursor="fuse: false; rayOrigin: mouse;" raycaster="near: 10; far: 10000; objects: .clickable"></a-camera>
-    `;
+  // Delete the original PDF file after conversion
+  fs.rm(filePath, { recursive: true, force: true }, (err) => {
+    if (err) console.error('Error deleting file:', err);
+  } );
 
-    // Add the arrows to the scene
-    gencode += `
-    <a-assets>
-      <img id="img1" src="http://localhost:3000/left-arrow.png" crossorigin="anonymous" />
-      <img id="img2" src="http://localhost:3000/right-arrow.png" crossorigin="anonymous" />`;
+  // Generate HTML code for the A-Frame scene
+  // This code creates a scene with a camera and an image target
+  let gencode = `
+  <a-scene id="example-target" mindar-image="imageTargetSrc: http://localhost:3000/${req.body.codetgt}.mind; uiScanning:yes; autoStart: false;" color-space="sRGB" renderer="colorManagement: true" vr-mode-ui="enabled: false" device-orientation-permission-ui="enabled: false">
+    <a-camera position="0 0 0" look-controls="enabled: false" cursor="fuse: false; rayOrigin: mouse;" raycaster="near: 10; far: 10000; objects: .clickable"></a-camera>
+  `;
 
-    // Get the first image and add it to the scene
-    gencode += `\n<img id="example-image-1" src="http://localhost:3000/${req.body.codetgt}.1.png" crossorigin="anonymous" />`;
+  // Add the arrows to the scene
+  gencode += `
+  <a-assets>
+    <img id="img1" src="http://localhost:3000/left-arrow.png" crossorigin="anonymous" />
+    <img id="img2" src="http://localhost:3000/right-arrow.png" crossorigin="anonymous" />`;
 
-    // Create the entity code for the images
-    // The entity code is a string that contains the HTML code for the images
-    let entitycode = ``;
+  // Get the first image and add it to the scene
+  gencode += `\n<img id="example-image-1" src="http://localhost:3000/${req.body.codetgt}.1.png" crossorigin="anonymous" />`;
 
-    // Add the first entity to the scene (the first image)
+  // Create the entity code for the images
+  // The entity code is a string that contains the HTML code for the images
+  let entitycode = ``;
+
+  // Add the first entity to the scene (the first image)
+  entitycode += `
+  <a-entity id="slide-1" class="slides">
+    <a-image src="#example-image-1"></a-image>
+    <a-image  class="clickable left-arrow" src="#img1" position="-0.7 0 0" height="0.15" width="0.15"></a-image>
+    <a-image  class="clickable right-arrow" src="#img2" position="0.7 0 0" height="0.15" width="0.15"></a-image>
+  </a-entity>`;
+
+  // Add the rest of the images to the scene
+  //for(let i = 2; i<=req.body.numpags; i++){
+  for(let i = 2; i < page; i++)
+  {
+    // Get the image and add it to the scene
+    gencode += `\n<img id="example-image-${i}" src="http://localhost:3000/${req.body.codetgt}.${i}.png" crossorigin="anonymous" />`;
+
+    // Add the entity for the image to the scene
+    // The entity is initially invisible and has the class "slides"
     entitycode += `
-    <a-entity id="slide-1" class="slides">
-      <a-image src="#example-image-1"></a-image>
-      <a-image  class="clickable left-arrow" src="#img1" position="-0.7 0 0" height="0.15" width="0.15"></a-image>
+    <a-entity id="slide-${i}" visible=false class="slides">
+      <a-image src="#example-image-${i}"></a-image>
+      <a-image  class="clickable left-arrow" src="#img1"  position="-0.7 0 0" height="0.15" width="0.15"></a-image>
       <a-image  class="clickable right-arrow" src="#img2" position="0.7 0 0" height="0.15" width="0.15"></a-image>
     </a-entity>`;
-    
-    // Add the rest of the images to the scene
-    //for(let i = 2; i<=req.body.numpags; i++){
-    for(let i = 2; i < page; i++)
-    {
-      // Get the image and add it to the scene
-      gencode += `\n<img id="example-image-${i}" src="http://localhost:3000/${req.body.codetgt}.${i}.png" crossorigin="anonymous" />`;
+  }
 
-      // Add the entity for the image to the scene
-      // The entity is initially invisible and has the class "slides"
-      entitycode += `
-      <a-entity id="slide-${i}" visible=false class="slides">
-        <a-image src="#example-image-${i}"></a-image>
-        <a-image  class="clickable left-arrow" src="#img1"  position="-0.7 0 0" height="0.15" width="0.15"></a-image>
-        <a-image  class="clickable right-arrow" src="#img2" position="0.7 0 0" height="0.15" width="0.15"></a-image>
-      </a-entity>`;
+  // Finally, add the assets to the scene and close the scene tag
+  gencode += `
+    \n</a-assets>
+    <a-entity id="slides-container" mindar-image-target="targetIndex: 0">`;
+
+  // Insert the entity code into the scene
+  gencode += entitycode;
+
+  // Close the entity and scene tags
+  gencode += `\n</a-entity></a-scene>`
+
+  // Write the generated HTML code to a file
+  // The file is saved in the public directory with the name given by the user
+  // The file is saved as a .html file
+  fs.writeFile(path.join(__dirname, 'public', `${req.body.codetgt}.html`), gencode, (err) => {
+    if (err) {
+      console.error('Error writing HTML file:', err);
+      return res.status(500).send('Failed to write HTML file.');
     }
-    
-    // Finally, add the assets to the scene and close the scene tag
-    gencode += `
-      \n</a-assets>
-      <a-entity id="slides-container" mindar-image-target="targetIndex: 0">`;
-
-    // Insert the entity code into the scene
-    gencode += entitycode;
-
-    // Close the entity and scene tags
-    gencode += `\n</a-entity></a-scene>`
-
-    // Write the generated HTML code to a file
-    // The file is saved in the public directory with the name given by the user
-    // The file is saved as a .html file
-    fs.writeFile(path.join(__dirname, 'public', `${req.body.codetgt}.html`), gencode, (err) => {
-      if (err) {
-        console.error('Error writing HTML file:', err);
-        return res.status(500).send('Failed to write HTML file.');
-      }
-      console.log(`HTML file created: ${req.body.codetgt}.html`);
-    } );
-    res.json({
-      success: true,
-      message: 'PDF converted successfully',
-    });
+    console.log(`HTML file created: ${req.body.codetgt}.html`);
+  } );
+  res.json({
+    success: true,
+    message: 'PDF converted successfully',
+  });
 });    
 
 //TODO: FILE UPLOADING RECIEVER
